Move field parsing into a separate class
