+++ OpenChart.Tests/src/Formats/StepMania/SM/TestSMSerializer.cs
@@ -16,69 +16,6 @@ namespace OpenChart.Tests.Formats.StepMania.SM
         }
 
         [Test]
-        public void Test_ParseDisplayBPM_Fixed()
-        {
-            var display = serializer.ParseDisplayBPM("120");
-            Assert.AreEqual(DisplayBPMType.Fixed, display.Type);
-            Assert.AreEqual(120, display.Min);
-        }
-
-        [Test]
-        public void Test_ParseDisplayBPM_Random()
-        {
-            var display = serializer.ParseDisplayBPM("*");
-            Assert.AreEqual(DisplayBPMType.Random, display.Type);
-        }
-
-        [Test]
-        public void Test_ParseDisplayBPM_Range()
-        {
-            var display = serializer.ParseDisplayBPM("60:120");
-            Assert.AreEqual(DisplayBPMType.Range, display.Type);
-            Assert.AreEqual(60, display.Min);
-            Assert.AreEqual(120, display.Max);
-        }
-
-        [TestCase("")]
-        [TestCase(" ")]
-        [TestCase(",")]
-        public void Test_ParseBPMs_Empty(string val)
-        {
-            var bpms = serializer.ParseBPMs(val);
-            Assert.AreEqual(0, bpms.Count);
-        }
-
-        [TestCase("100=0", 100, 0)]
-        [TestCase("123.45=1.2345", 123.45, 1.2345)]
-        [TestCase("0=-1", 0, -1)]
-        public void Test_ParseBPMs_Single(string val, double beat, double bpm)
-        {
-            var bpms = serializer.ParseBPMs(val);
-            Assert.AreEqual(1, bpms.Count);
-            Assert.AreEqual(beat, bpms[0].Beat);
-            Assert.AreEqual(bpm, bpms[0].Value);
-        }
-
-        [TestCase("")]
-        [TestCase(" ")]
-        [TestCase(",")]
-        public void Test_ParseStops_Empty(string val)
-        {
-            var stops = serializer.ParseStops(val);
-            Assert.AreEqual(0, stops.Count);
-        }
-
-        [TestCase("100=0", 100, 0)]
-        [TestCase("123.45=1.2345", 123.45, 1.2345)]
-        public void Test_ParseStops_Single(string val, double beat, double seconds)
-        {
-            var stops = serializer.ParseStops(val);
-            Assert.AreEqual(1, stops.Count);
-            Assert.AreEqual(beat, stops[0].Beat);
-            Assert.AreEqual(seconds, stops[0].Seconds);
-        }
-
-        [Test]
         public void Test_Deserialize_NoCharts()
         {
             var str = @"
